rename: move files and fix filename
